+++ packages/auth/lib/cognitoClient.ts
@@ -7,6 +7,7 @@ import {
 } from '@aws-sdk/client-cognito-identity-provider'
 import { type User } from 'next-auth'
 import invariant from 'tiny-invariant'
+import { z } from 'zod'
 
 import { createHmac } from 'crypto'
 
@@ -15,8 +16,14 @@ import { prisma } from '@weareinreach/db'
 
 import { decodeCognitoIdJwt } from './cognitoJwt'
 import { generateUserSession } from './genUserSession'
-import { CognitoSessionSchema } from './userLogin'
 
+export const CognitoSessionSchema = z.object({
+	AccessToken: z.string(),
+	ExpiresIn: z.number(),
+	IdToken: z.string(),
+	RefreshToken: z.string().optional(),
+	TokenType: z.string(),
+})
 export const cognito = new CognitoIdentityProvider({
 	credentials: {
 		accessKeyId: getEnv('COGNITO_ACCESS_KEY'),
